# docs/swagger.yaml
basePath: /
definitions:
  echo.HTTPError:
    properties:
      message: {}
    type: object
  github_com_simesaba80_toybox-back_internal_interface_schema.AssetResponse:
    properties:
      asset_type:
        type: string
      created_at:
        type: string
      extension:
        type: string
      id:
        type: string
      updated_at:
        type: string
      url:
        type: string
      user_id:
        type: string
      work_id:
        type: string
    type: object
  github_com_simesaba80_toybox-back_internal_interface_schema.CollaboratorResponse:
    properties:
      avatar_url:
        type: string
      display_name:
        type: string
      id:
        type: string
    type: object
...
  github_com_simesaba80_toybox-back_internal_interface_schema.CreateWorkInput:
    properties:
      asset_ids:
        items:
          type: string
        type: array
      collaborator_ids:
        items:
          type: string
        type: array
      description:
        type: string
      tag_ids:
        items:
          type: string
        type: array
      thumbnail_asset_id:
        type: string
      title:
        maxLength: 100
        type: string
      urls:
        items:
          type: string
        type: array
      visibility:
        enum:
        - public
        - private
        - draft
        type: string
    required:
    - asset_ids
    - description
    - tag_ids
    - thumbnail_asset_id
    - title
    - visibility
    type: object
  github_com_simesaba80_toybox-back_internal_interface_schema.CreateWorkOutput:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
      visibility:
        type: string
    type: object
...
  github_com_simesaba80_toybox-back_internal_interface_schema.GetWorkOutput:
    properties:
      assets:
        items:
          $ref: '#/definitions/github_com_simesaba80_toybox-back_internal_interface_schema.AssetResponse'
        type: array
      collaborators:
        items:
          $ref: '#/definitions/github_com_simesaba80_toybox-back_internal_interface_schema.CollaboratorResponse'
        type: array
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_favorite:
        type: boolean
      tags:
        items:
          $ref: '#/definitions/github_com_simesaba80_toybox-back_internal_interface_schema.TagResponse'
        type: array
      thumbnail_asset_id:
        type: string
      thumbnail_url:
        type: string
      title:
        type: string
      updated_at:
        type: string
      urls:
        items:
          type: string
        type: array
      user:
        $ref: '#/definitions/github_com_simesaba80_toybox-back_internal_interface_schema.UserInWorkResponse'
      visibility:
        type: string
    type: object
...
  github_com_simesaba80_toybox-back_internal_interface_schema.LegacyProxyErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  github_com_simesaba80_toybox-back_internal_interface_schema.TagResponse:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  github_com_simesaba80_toybox-back_internal_interface_schema.UpdateWorkInput:
    properties:
      asset_ids:
        items:
          type: string
        minItems: 1
        type: array
      collaborator_ids:
        items:
          type: string
        type: array
      description:
        type: string
      tag_ids:
        items:
          type: string
        minItems: 1
        type: array
      thumbnail_asset_id:
        type: string
      title:
        maxLength: 100
        type: string
      urls:
        items:
          type: string
        type: array
      visibility:
        enum:
        - public
        - private
        - draft
        type: string
    type: object
...
  github_com_simesaba80_toybox-back_internal_interface_schema.UserInWorkResponse:
    properties:
      avatar_url:
        type: string
      display_name:
        type: string
      id:
        type: string
    type: object
...
  github_com_simesaba80_toybox-back_internal_interface_schema.WorkListResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total_count:
        type: integer
      works:
        items:
          $ref: '#/definitions/github_com_simesaba80_toybox-back_internal_interface_schema.GetWorkOutput'
        type: array
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is the API server for the Toybox application.
  title: Toybox API
  version: "1.0"
paths:
...
  /api/v1/works:
    get:
      description: Temporary read-only pass-through to the legacy ToyBox API. The
        response schema is owned by the legacy service.
      parameters:
      - description: Maximum number of works
        in: query
        name: limit
        type: integer
      - description: Visibility filter
        in: query
        name: visibility
        type: string
      - description: Oldest work cursor
        in: query
        name: oldest_work_id
        type: string
      - description: Newest work cursor
        in: query
        name: newest_work_id
        type: string
      - description: Comma-separated tag names
        in: query
        name: tag_names
        type: string
      - description: Comma-separated tag IDs
        in: query
        name: tag_ids
        type: string
      - description: Search text
        in: query
        name: search_word
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/github_com_simesaba80_toybox-back_internal_interface_schema.LegacyProxyErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/github_com_simesaba80_toybox-back_internal_interface_schema.LegacyProxyErrorResponse'
      summary: Get works from the legacy API (v1)
      tags:
      - legacy-proxy
  /api/v2/works:
    get:
      description: Temporary read-only pass-through to the legacy ToyBox API. The
        response schema is owned by the legacy service.
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Maximum number of works
        in: query
        name: limit
        type: integer
      - description: Visibility filter
        in: query
        name: visibility
        type: string
      - description: Comma-separated tag names
        in: query
        name: tag_names
        type: string
      - description: Comma-separated tag IDs
        in: query
        name: tag_ids
        type: string
      - description: Search text
        in: query
        name: search_word
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/github_com_simesaba80_toybox-back_internal_interface_schema.LegacyProxyErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/github_com_simesaba80_toybox-back_internal_interface_schema.LegacyProxyErrorResponse'
      summary: Get works from the legacy API (v2)
      tags:
      - legacy-proxy
...
  /auth/works:
    post:
      consumes:
      - application/json
      description: Create a new work with the input payload
      parameters:
      - description: Work to create
        in: body
        name: work
        required: true
        schema:
          $ref: '#/definitions/github_com_simesaba80_toybox-back_internal_interface_schema.CreateWorkInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_simesaba80_toybox-back_internal_interface_schema.CreateWorkOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - BearerAuth: []
      summary: Create a new work
      tags:
      - works
...
  /auth/works/{work_id}:
    delete:
      description: Delete a work by ID (only owner can delete)
      parameters:
      - description: Work ID
        in: path
        name: work_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - BearerAuth: []
      summary: Delete a work
      tags:
      - works
    patch:
      consumes:
      - application/json
      description: Update a work by ID (only owner can update)
      parameters:
      - description: Work ID
        in: path
        name: work_id
        required: true
        type: string
      - description: Work to update
        in: body
        name: work
        required: true
        schema:
          $ref: '#/definitions/github_com_simesaba80_toybox-back_internal_interface_schema.UpdateWorkInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_simesaba80_toybox-back_internal_interface_schema.GetWorkOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - BearerAuth: []
      summary: Update a work
      tags:
      - works
...
  /works:
    get:
      description: Get all works with pagination and optional tag filter (OR search)
      parameters:
      - description: 'Limit per page (default: 20, max: 100)'
        in: query
        name: limit
        type: integer
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: Comma-separated tag IDs for filtering (OR search)
        in: query
        name: tag_ids
        type: string
      - description: 'Sort order: newest (default) or oldest'
        in: query
        name: sort
        type: string
      - description: 'Filter by visibility: public or private'
        in: query
        name: visibility
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_simesaba80_toybox-back_internal_interface_schema.WorkListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - BearerAuth: []
      summary: Get all works
      tags:
      - works
...
  /works/{work_id}:
    get:
      description: Get a work by ID
      parameters:
      - description: Work ID
        in: path
        name: work_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_simesaba80_toybox-back_internal_interface_schema.GetWorkOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - BearerAuth: []
      summary: Get a work by ID
      tags:
      - works
...
  /works/users/{user_id}:
    get:
      description: Get works by user ID
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: 'Limit per page (default: 20, max: 100)'
        in: query
        name: limit
        type: integer
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_simesaba80_toybox-back_internal_interface_schema.WorkListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - BearerAuth: []
      summary: Get works by user ID
      tags:
      - works
securityDefinitions:
  BearerAuth:
    description: Bearer token for authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"